corrections avec Luc

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dasouzda <dasouzda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 15:26:07 by dasouzda          #+#    #+#             */
-/*   Updated: 2024/12/16 16:39:27 by dasouzda         ###   ########.fr       */
+/*   Updated: 2025/01/02 12:57:42 by dasouzda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		ft_printf(const char *format, ...);
 void	ft_putchar_count(char c, size_t *count);
 void	ft_putstr_count(char *str, size_t *count);
 void	ft_putptr_count(void *ptr, size_t *count);
-void	ft_putnbr_base_count(int nb, char *base, size_t *count);
+void	ft_putnbr_base_count(unsigned long nb, char *base, size_t *count);
 void	ft_putnbr_count(int n, size_t *count);
 void	ft_putunsigned_count(unsigned int n, size_t *count);
